% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sifter.R
\name{sifter}
\alias{sifter}
\title{The main function for sifting algorithm}
\usage{
sifter(
  data_main,
  data_art = NULL,
  class_col,
  internal_number_col,
  num_of_classes = NULL,
  cutting_at_height = NULL,
  labels_cex = 0.7,
  mai_bottom = 1.5,
  clust_algo = "hc",
  dist_method = "euclidean",
  aggl_method = "ward",
  plot_2D = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data_main}{Input data. See \code{"toy_example.csv"} file available through the \code{system.file()} function.}

\item{data_art}{Input data with artefacts. See \code{"toy_example_artefacts.csv"} file available through the \code{system.file()} function.}

\item{class_col}{The column number in the input data that contains the class identifier.}

\item{internal_number_col}{The column number in the input data that contains unique IDs for each record.}

\item{num_of_classes}{The number of classes into which the tree should be divided.}

\item{cutting_at_height}{The height at which the tree is cut.}

\item{labels_cex}{\code{cex} param for tree leaves/}

\item{mai_bottom}{\code{mai} bottom param. Used when leaf labels are long and may not fit entirely in the plot.}

\item{clust_algo}{One of these: \code{"hc"} - the standard hierarchical clustering, \code{"agnes"} - an enhanced agglomerative hierarchical clustering algorithm, \code{"diana"} - a divisive (top-down) hierarchical clustering algorithm.}

\item{dist_method}{One of these: \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.}

\item{aggl_method}{For \code{clust_algo == "hc"} - \code{"ward"}, \code{"single"}, \code{"complete"}, \code{"average"}, \code{"mcquitty"}, \code{"median"}, \code{"centroid"}.For \code{clust_algo == "agnes"} - \code{"ward"}, \code{"single"}, \code{"complete"}, \code{"average"}, \code{"weighted"}, \code{"flexible"}, \code{"gaverage"}.}

\item{plot_2D}{If \code{TRUE} 2D cluster plot  will be displayed on the screen.}

\item{verbose}{If \code{TRUE} results will be displayed on the screen (AHC tree and 2D cluster plot).}
}
\value{
TODO
}
\description{
The function generates a series of Agglomerative Hierarchical Clustering (AHC) trees.
}
\examples{
file_toy <- system.file(
 "extdata", "toy_example.csv", package = "sifter")

file_toy_artefacts <- system.file(
  "extdata", "toy_example_artefacts.csv", package = "sifter")

data <- read.csv(
  file = file_toy,
  dec = ".",
  sep = "\t",
  header = TRUE,
  stringsAsFactors = TRUE
)

data_artefact <- read.csv(
  file = file_toy_artefacts,
  dec = ".",
  sep = "\t",
  header = TRUE,
  stringsAsFactors = TRUE
)

print(data)
print(data_artefact)

# Only the required parameters.
out <- sifter(
  data_main = data,
  class_col = 5,
  internal_number_col = 6
)

# Specify the number of classes.
out <- sifter(
  data_main = data,
  class_col = 5,
  internal_number_col = 6,
  num_of_classes = 2
)

# Specifying the height at which we cut the tree.
out <- sifter(
  data_main = data,
  class_col = 5,
  internal_number_col = 6,
  cutting_at_height = 8
)

# Specifying the height at which we cut the tree AND an artefact.
out <- sifter(
  data_main = data,
  data_art = data_artefact[2,],
  class_col = 5,
  internal_number_col = 6,
  cutting_at_height = 8
)

# Tree and 2D cluster plot.
out <- sifter(
  data_main = data,
  class_col = 5,
  internal_number_col = 6,
  num_of_classes = 2,
  data_art = data_artefact[2,],
  plot_2D = TRUE
)


}
