% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sifter_toolbox_interactive_mode.R
\name{sifter_toolbox_interactive_mode}
\alias{sifter_toolbox_interactive_mode}
\title{A wrapper for the sifter_toolbox() function}
\usage{
sifter_toolbox_interactive_mode(
  data_main,
  data_art,
  class_col,
  internal_number_col,
  num_of_classes = NULL,
  class_to_highlight = NULL,
  labels_cex = 0.7,
  mai_bottom = 1.5,
  clust_algo = "hc",
  dist_method = "euclidean",
  aggl_method = "ward",
  dir = "figs_sifter",
  save_2D = FALSE
)
}
\arguments{
\item{data_main}{TODO}

\item{data_art}{TODO}

\item{class_col}{TODO}

\item{internal_number_col}{TODO}

\item{num_of_classes}{TODO}

\item{class_to_highlight}{TODO}

\item{labels_cex}{TODO}

\item{mai_bottom}{TODO}

\item{clust_algo}{TODO}

\item{dist_method}{TODO}

\item{aggl_method}{TODO}

\item{dir}{TODO}

\item{save_2D}{TODO}
}
\value{
TODO
}
\description{
A wrapper for the sifter_toolbox() function
}
\examples{
if (interactive()) {

file_toy <- system.file(
 "extdata", "toy_example.csv", package = "sifter")

file_toy_artefacts <- system.file(
  "extdata", "toy_example_artefacts.csv", package = "sifter")

data <- read.csv(
  file = file_toy,
  dec = ".",
  sep = "\t",
  header = TRUE,
  stringsAsFactors = TRUE
)

data_artefact <- read.csv(
  file = file_toy_artefacts,
  dec = ".",
  sep = "\t",
  header = TRUE,
  stringsAsFactors = TRUE
)

sifter_toolbox_interactive_mode(
  data_main = data,
  data_art = data_artefact[2,],
  class_col = 5,
  internal_number_col = 6,
  num_of_classes = 2,
  dir = "sifter_results",
  save_2D = FALSE
)

}

}
