% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sifter.R
\name{sifter}
\alias{sifter}
\title{The main function for sifting algorithm which generats a series of Agglomerative Hierarchical Clustering (AHC)trees}
\usage{
sifter(
  data_main,
  data_art = NULL,
  class_col,
  internal_number_col,
  num_of_classes = NULL,
  cutting_at_height = NULL,
  labels_cex = 0.7,
  mai_bottom = 1.5,
  clust_algo = "hc",
  dist_method = "euclidean",
  aggl_method = "ward",
  plot_2D = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data_main}{TODO}

\item{data_art}{TODO}

\item{class_col}{TODO}

\item{internal_number_col}{TODO}

\item{num_of_classes}{TODO}

\item{cutting_at_height}{TODO}

\item{labels_cex}{TODO}

\item{mai_bottom}{TODO}

\item{clust_algo}{TODO}

\item{dist_method}{TODO}

\item{aggl_method}{TODO}

\item{plot_2D}{TODO}

\item{verbose}{TODO}
}
\value{
TODO
}
\description{
The main function for sifting algorithm which generats a series of Agglomerative Hierarchical Clustering (AHC)trees
}
\examples{
file_toy <- system.file("extdata", "toy_example.csv", package = "sifter")
file_toy_artefacts <- system.file("extdata", "toy_example_artefacts.csv", package = "sifter")

data <- read.csv(file = file_toy, dec = ".", sep = "\t", header = T, stringsAsFactors = T)
data_artefact <- read.csv(file = file_toy_artefacts, dec = ".", sep = "\t", header = T, stringsAsFactors = T)

out <- sifter(
  data_main = data,
  class_col = 5,
  internal_number_col = 6,
  verbose = TRUE
)

out <- sifter(
  data_main = data,
  class_col = 5,
  internal_number_col = 6,
  num_of_classes = 2,
  verbose = TRUE
)

out <- sifter(
  data_main = data,
  class_col = 5,
  internal_number_col = 6,
  cutting_at_height = 8,
  verbose = TRUE
)

out <- sifter(
  data_main = data,
  data_art = data_artefact[2,],
  class_col = 5,
  internal_number_col = 6,
  cutting_at_height = 8,
  verbose = TRUE
)

}
