% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sifter_toolbox.R
\name{sifter_toolbox}
\alias{sifter_toolbox}
\title{A helper function that automatically calls \code{sifter()} function in a loop}
\usage{
sifter_toolbox(
  data_main,
  data_art,
  class_col,
  internal_number_col,
  num_of_classes,
  labels_cex = 0.7,
  mai_bottom = 1.5,
  clust_algo = "hc",
  dist_method = "euclidean",
  aggl_method = "ward",
  dir = "figs_sifter",
  save_2D = FALSE
)
}
\arguments{
\item{data_main}{Same as in \code{sifter()} function.}

\item{data_art}{Same as in \code{sifter()} function.}

\item{class_col}{Same as in \code{sifter()} function.}

\item{internal_number_col}{Same as in \code{sifter()} function.}

\item{num_of_classes}{Same as in \code{sifter()} function.}

\item{labels_cex}{Same as in \code{sifter()} function.}

\item{mai_bottom}{Same as in \code{sifter()} function.}

\item{clust_algo}{Same as in \code{sifter()} function.}

\item{dist_method}{Same as in \code{sifter()} function.}

\item{aggl_method}{Same as in \code{sifter()} function.}

\item{dir}{The directory where the results will be saved. If the directory
does not exist, it is first created.}

\item{save_2D}{Logical. If \code{TRUE}, the 2D cluster plot is saved to
disk as a \code{png} file.}
}
\value{
The function generates results and saves them to the directory specified
by the \code{dir} parameter. If the directory doesn't already exist, it is
created. Images containing the trees created in subsequent steps are
saved there, until a homogeneous class is achieved. If \code{save_2D == TRUE},
the 2D clusters constructed for each tree are also saved.
}
\description{
A helper function that automatically calls \code{sifter()} function in a loop
}
\examples{
if (interactive()) {

# Reading a very simple, small toy dataset.
file_toy <- system.file(
 "extdata", "toy_example.csv", package = "sifter")

file_toy_artefacts <- system.file(
  "extdata", "toy_example_artefacts.csv", package = "sifter")

data <- read.csv(
  file = file_toy,
  dec = ".",
  sep = "\t",
  header = TRUE,
  stringsAsFactors = TRUE
)

data_artefact <- read.csv(
  file = file_toy_artefacts,
  dec = ".",
  sep = "\t",
  header = TRUE,
  stringsAsFactors = TRUE
)

sifter_toolbox(
  data_main = data,
  data_art = data_artefact[2,],
  class_col = 5,
  internal_number_col = 6,
  num_of_classes = 2,
  dir = "sifter_results",
  save_2D = TRUE
)

}

}
